refactor(powiazania_autorow): wydziel warstwę danych do queries.py
views.py urósł do 653 linii (>600) przez warstwę filtra. Przeniosłem
helpery danych, klasę Filtr, context manager _limit_czasu i stałe do
nowego queries.py (447 lin.); views.py zostają tylko 4 klasy widoków
(225 lin.). Czysto mechaniczne 1:1, bez zmiany logiki/sygnatur/URL-i.
